refactor(common): step2 패키지명 리팩토링
