+++ src/main/kotlin/calculator/calculator/Calculator.kt
@@ -1,4 +1,4 @@
-package step2.calculator
+package calculator.calculator
 
 interface Calculator {
     /**
