+++ src/main/kotlin/calculator/util/Util.kt
@@ -1,4 +1,4 @@
-package step2.util
+package calculator.util
 
 fun Int.isOdd() = this % 2 != 0
